ml/requirements.txt
opencv-python
Pillow
numpy
pyserial
opencv-contrib-python